style: clear the mechanical part of the advisory lint backlog
golangci-lint is advisory here and reports the whole repository, and
GitHub shows only the first ten annotations per job, so every PR carries
the same ten pre-existing warnings. This clears everything mechanical:
gofmt / gofumpt / goimports on 40 files (mostly stdlib imports that go
fix added outside the stdlib group), checked type assertions in tests,
the dead err == nil branch after LavaFormatTrace in the parser (it never
returns nil), slices.Concat for the appendAssign in spec.go, dedicated
paragraphs for Deprecated notices, an ineffectual assignment, three
unnecessary conversions, and explained nolint directives on the two
mocks that return nil, nil by design and the test that exercises a typed
nil on purpose.

Two files keep golangci-lint's gofmt form for a multi-value return of
composite literals, which Go 1.27's gofmt indents differently; nothing
in CI runs raw gofmt, and the linter is what gates the annotation.

Left alone, 68 issues that change behaviour or need a migration: ST1005
(rewording error strings), ST1012 (renaming exported error variables),
SA1019 (deprecated protoreflect / golang-protobuf / h2c packages).

+++ protocol/lavasession/endpoint_probe_reenable_test.go
@@ -49,7 +49,7 @@ func TestRecordProbeVerdict_ReEnablesAfterKDistinctPostDisablePolls(t *testing.T
 	require.True(t, e.Enabled)
 
 	e.mu.RLock()
-	require.Equal(t, uint64(MaxConsecutiveConnectionAttempts-probeReenableTrialBudget), e.ConnectionRefusals,
+	require.Equal(t, MaxConsecutiveConnectionAttempts-probeReenableTrialBudget, e.ConnectionRefusals,
 		"a probe-granted re-enable carries only a TRIAL refusal budget, not the full threshold (MAG-2550)")
 	require.Equal(t, uint64(0), e.consecutiveHealthyProbes, "the hysteresis streak resets after re-enable")
 	require.True(t, e.disabledAt.IsZero(), "disabledAt cleared on re-enable")
